requirements.txt
# 基本依赖
pillow>=9.0.0
numpy>=1.20.0
requests>=2.25.0

# 可选依赖 (取消注释以安装)
# torch>=1.10.0  # 用于PyTorch张量支持
# pillow_heif>=0.4.0  # 用于HEIF/HEIC格式支持 